+++ R/get.R
@@ -1,11 +1,12 @@
 #' Get endpoint data
 #' 
 #' @param url shortcut API url.  Defaults to API V2
-#' @param endpoint  data to list. One of "categories", "epic-workflow", "epics", 
-#' "files", "labels", "linked-files", "members", "milestones", "projects", 
-#' "repositories", "teams", "workflows", "iterations"
+#' @param endpoint  data to list. One of "categories", "entity-templates","epic-workflow", "epics"
+#' ,"files", "groups", "iterations","labels", "linked-files", "members",
+#' "milestones", "projects", "repositories", 
+#' "workflows"
 #' @param config additional configuration to add to header
-#' @param ch_token shortcut API token. 
+#' @param sc_token shortcut API token. 
 #' @param response_type either "full" (all data) or "minimal" (just "entity_type", "id", 
 #'"name", "description"). Defaults to "full"
 #'
@@ -14,30 +15,30 @@
 #' \dontrun{
 #'
 #' # Retrieve all projects
-#' ch_list_all(endpoint = "projects")
+#' sc_list_all(endpoint = "projects")
 #' 
 #' # List epics then retrieve full details about the first one
-#' df <- ch_list_all("epics", response_type = "minimal")
-#' ch_get_one("epics", id = df[ 1, "id"] )
+#' df <- sc_list_all("epics", response_type = "minimal")
+#' sc_get_one("epics", id = df[ 1, "id"] )
 #' }
 #' 
 #' @export
-ch_list_all <- function(endpoint = NULL,
+sc_list_all <- function(endpoint = NULL,
                         url = NULL,
                         config=list(),
-                        ch_token = get_token(),
+                        sc_token = get_token(),
                         response_type = "full"){
   
-  match.arg(endpoint, choices = c("categories", "epic-workflow", "epics", 
-            "files", "labels", "linked-files", "members",
+  match.arg(endpoint, choices = c("categories", "entity-templates","epic-workflow", "epics", 
+            "files", "groups", "iterations","labels", "linked-files", "members",
             "milestones", "projects", "repositories", 
-            "teams", "workflows", "iterations"))
+             "workflows"))
   match.arg(response_type, choices = c("full", "minimal"))
   if(is.null(endpoint) & is.null(url)) stop("Please specify one of url or endpoint")
   if(!is.null(endpoint) & !is.null(url)) stop("Please specify only one of full url or endpoint")
 
-  df <- ch_GET(url = ifelse( is.null(url), 
-                             ch_url(endpoint = endpoint, ch_token = ch_token), 
+  df <- sc_GET(url = ifelse( is.null(url), 
+                             sc_url(endpoint = endpoint, sc_token = sc_token), 
                              url),
                config = config)
 
@@ -59,45 +60,47 @@ ch_list_all <- function(endpoint = NULL,
 #' Get endpoint data for one item
 #' 
 #' @param url shortcut API url.  Defaults to API V2
-#' @param endpoint  data to list. One of "categories",  "epics", 
-#' "files", "labels", "linked-files",  "milestones", "projects", "repositories",
-#' "teams", "iterations"
+#' @param endpoint  data to list. One of "categories", "entity-templates", "epic-workflows", "epics", 
+#' "files", "groups", "iterations"  ,"labels", "linked-files", 
+#' "members", "milestones", "projects", "repositories", "stories",
+#' "story-links"
 #' @param id the id for the record you want to retrieve
 #' @param config additional configuration to add to header
-#' @param ch_token shortcut API token. 
+#' @param sc_token shortcut API token. 
 #'
 #'@examples
 #'\dontrun{
 #' # Retrieve all projects
-#' ch_list_all(endpoint = "projects")
+#' sc_list_all(endpoint = "projects")
 #' 
 #' # List epics then retrieve full details about the first one
-#' df <- ch_list_all("epics", response_type = "minimal")
-#' ch_get_one("epics", id = df[ 1, "id"] )
+#' df <- sc_list_all("epics", response_type = "minimal")
+#' sc_get_one("epics", id = df[ 1, "id"] )
 #' }
 #' 
 #' @export
-ch_get_one <- function( id, 
+sc_get_one <- function( id, 
                         endpoint = NULL,
                         url = NULL,
                         config=list(),
-                        ch_token = get_token()){
+                        sc_token = get_token()){
   
   # TODO: "members" endpoint is failing - need to add
   
-  match.arg(endpoint, choices = c("categories","epics", 
-                                  "files", "labels", "linked-files", 
-                                  "milestones", "projects", "repositories", "stories",
-                                  "story-links", "teams", "iterations"))
+  match.arg(endpoint, choices = c("categories", "entity-templates", "epic-workflows", "epics", 
+                                  "files", "groups", "iterations"  ,"labels", "linked-files", 
+                                  "members", "milestones", "projects", "repositories", "stories",
+                                  "story-links"))
   
   if(is.null(endpoint) & is.null(url)) stop("Please specify one of url or endpoint")
   if(!is.null(endpoint) & !is.null(url)) stop("Please specify only one of full url or endpoint")
   
-  df <- ch_GET(url = ifelse( is.null(url), 
-                             ch_url(endpoint = endpoint, id = id, ch_token = ch_token), 
+  df <- sc_GET(url = ifelse( is.null(url), 
+                             sc_url(endpoint = endpoint, id = id, sc_token = sc_token), 
                              url),
                config = config)
   
   return(df)
 
 }
+