+++ R/search.R
@@ -2,38 +2,38 @@
 #' Search stories
 #'
 #' @param query Search query.  See https://help.shortcut.io/hc/en-us/articles/115005967026 for more detail on allowed search structure.
-#' @param search_limit Maximum records to retrieve.  Defaults to 1000.
+#' @param searsc_limit Maximum records to retrieve.  Defaults to 1000.
 #' @param url Alternatively specify query as a url string. Useful if you want to use a different version of the API (default is V2)
 #' @param config additional configuration to add to header
 #' @param response_type either "full" (all data as a nested list) or 
 #' "summary" (flattened data frame with nested lists removed) or 
 #' "minimal" (just "entity_type", "id", "name", "description"). Defaults to "summary"
-#' @param ch_token shortcut API token.
+#' @param sc_token shortcut API token.
 #'
 #'
 #'@examples
 #'\dontrun{
 #' # Search for the term "foo" anywhere in stories
-#' ch_search_stories(query = "foo")
+#' sc_searsc_stories(query = "foo")
 #'
 #' # Search for epic names that contain specified text
 #' # Note - no whitespace areound the ":"
-#' ch_search_stories(query = "epic:foo")
+#' sc_searsc_stories(query = "epic:foo")
 #'
 #' # Combine search terms with "AND" ("OR" combinations are not supported by the API V2)
 #' # Note the whitespace between each filter
-#' ch_search_stories(query = "epic:foo is:done type:bug")
+#' sc_searsc_stories(query = "epic:foo is:done type:bug")
 #'
 #'
 #' }
 #'
 #' @export
-ch_search_stories <- function( query,
-                        search_limit = 1000,
+sc_searsc_stories <- function( query,
+                        searsc_limit = 1000,
                         url = NULL,
                         config=list(),
                         response_type = "summary",
-                        ch_token = get_token()){
+                        sc_token = get_token()){
 
 
   match.arg(response_type, c("full", "summary", "minimal"))
@@ -43,11 +43,11 @@ ch_search_stories <- function( query,
   
   if(is.null(url)){ 
 
-    url <- ch_url(endpoint = "search/stories", 
-                  query = query, ch_token = ch_token) 
+    url <- sc_url(endpoint = "search/stories", 
+                  query = query, sc_token = sc_token) 
   }
   
-  res <- ch_GET(url =  url, config = config)
+  res <- sc_GET(url =  url, config = config)
   
   
   if( length(res[["data"]]) == 0){
@@ -60,11 +60,11 @@ ch_search_stories <- function( query,
   
   nrec <- nrow(lst[[1]]) 
   i <- 2
-  while( length(res[["next"]]) == 1 & nrec < search_limit){
+  while( length(res[["next"]]) == 1 & nrec < searsc_limit){
     
     url <- paste0( "https://api.shortcut.io", res[["next"]], 
-                   "&token=", ch_token)
-    res <- ch_GET(url =  url, config = config)
+                   "&token=", sc_token)
+    res <- sc_GET(url =  url, config = config)
     lst[[i]] <- as.data.frame( res$data, row.names = NULL  )
     
     nrec <- nrec+nrow(lst[[i]])
@@ -84,8 +84,8 @@ ch_search_stories <- function( query,
                       'created_at', 'moved_at', 'lead_time', 'cycle_time')
             
             df <- do.call("rbind", lapply(lst, "[", cols) ) 
-            if(nrow(df) > search_limit){
-              return(df[1:search_limit,])
+            if(nrow(df) > searsc_limit){
+              return(df[1:searsc_limit,])
             } else{
             return(df)
             }
@@ -94,8 +94,8 @@ ch_search_stories <- function( query,
             
             cols <- c("entity_type", "id", "name", "description")
             df <- do.call("rbind", lapply(lst, "[", cols) ) 
-            if(nrow(df) > search_limit){
-              return(df[1:search_limit,])
+            if(nrow(df) > searsc_limit){
+              return(df[1:searsc_limit,])
             } else{
               return(df)
             }