+++ tests/testthat/test-api-settings.R
@@ -5,7 +5,7 @@ test_that("API token is set", {
   reset_api()
   set_token("mytoken")
   expect_true( 
-    getOption("shortcutr")[["ch_token"]] == "mytoken"
+    getOption("shortcutr")[["sc_token"]] == "mytoken"
     )
 })
 
@@ -14,7 +14,7 @@ test_that("url is changed with set_url", {
   reset_api()
   set_url("myurl")
   expect_true( 
-    getOption("shortcutr")[["ch_base_url"]] == "myurl"
+    getOption("shortcutr")[["sc_base_url"]] == "myurl"
   )
 })
 
@@ -42,7 +42,7 @@ test_that("default url is set", {
   
   reset_api()
   expect_true( 
-    getOption("shortcutr")[["ch_base_url"]] == "https://api.shortcut.io/api/v2"
+    getOption("shortcutr")[["sc_base_url"]] == "https://api.app.shortcut.com/api/v3"
   )
 })
 
@@ -50,6 +50,7 @@ test_that("default token is NA", {
   
   reset_api()
   expect_true(
-    is.na(getOption("shortcutr")[["ch_token"]] )
+    is.na(getOption("shortcutr")[["sc_token"]] )
   )
 })
+